bug fixed: compatibility for 64bit machines
git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@8235 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10

+++ src/c++/common/spds/EmailData.cpp
@@ -46,7 +46,7 @@ EmailData::EmailData()
 int EmailData::parse(const BCHAR *syncmlData, size_t len)
 {
     int ret = 0;
-    size_t start, end, size;        
+    unsigned int start, end, size;        
     StringBuffer* s = new StringBuffer(syncmlData, len);
     size = s->length();
     // FIXME: remove these replace once the server has fixed the message.
